backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
SQLAlchemy>=2.0
pydantic